requirements.txt
apache-airflow
requests~=2.31.0